fix: Styling changes
