+++ dev-client/src/components/common/CheckboxGroup.tsx
@@ -1,4 +1,4 @@
-import {Box, FormControl, HStack} from 'native-base';
+import {Box, FormControl, HStack, VStack} from 'native-base';
 import {useCallback, useMemo} from 'react';
 import {useTranslation} from 'react-i18next';
 import CheckBox from '@react-native-community/checkbox';
@@ -42,22 +42,24 @@ const CheckboxGroup = ({
           onValueChange={onSelectAll}
           value={selectAllChecked}
         />
-        <FormControl.Label htmlFor={'select-all-' + groupName}>
+        <FormControl.Label htmlFor={'select-all-' + groupName} variant="body1">
           {t('general.select_all')}
         </FormControl.Label>
       </HStack>
-      {checkboxes.map(({label, id, checked}) => (
-        <HStack key={id}>
-          <CheckBox
-            id={'checkbox-' + id}
-            onValueChange={onChangeValue(groupId, id)}
-            value={checked}
-          />
-          <FormControl.Label htmlFor={'checkbox-' + id}>
-            {label}
-          </FormControl.Label>
-        </HStack>
-      ))}
+      <VStack px="20px">
+        {checkboxes.map(({label, id, checked}) => (
+          <HStack key={id}>
+            <CheckBox
+              id={'checkbox-' + id}
+              onValueChange={onChangeValue(groupId, id)}
+              value={checked}
+            />
+            <FormControl.Label htmlFor={'checkbox-' + id} variant="body1">
+              {label}
+            </FormControl.Label>
+          </HStack>
+        ))}
+      </VStack>
     </Box>
   );
 };
